+++ lib/rus3/evaluator.rb
@@ -14,6 +14,7 @@ module Rus3
     class Environment
       include Rus3::Procedure::Control
       include Rus3::Procedure::Write
+      include Rus3::Procedure::Vector
       include Rus3::Procedure::List
       include Rus3::Procedure::Predicate
       include Rus3::EmptyList